[CHANGE] make survey-list as server component
